requirements.txt
qdrant-client
openai
pandas
numpy
python-dotenv
tqdm